+++ src/lightly_train/_commands/train_task_helpers.py
@@ -195,25 +195,53 @@ def pretty_format_args_dict(args: dict[str, Any]) -> dict[str, Any]:
 
 def get_transform_args(
     train_model_cls: type[TrainModel],
+    transform_args: dict[str, Any] | None,
     ignore_index: int | None,
 ) -> tuple[TaskTransformArgs, TaskTransformArgs]:
     if train_model_cls.task != "semantic_segmentation" and ignore_index is not None:
         raise ValueError(
             "`ignore_index` is only supported for semantic segmentation tasks."
         )
+    transform_args = {} if transform_args is None else transform_args.copy()
+    if ignore_index is not None:
+        transform_args["ignore_index"] = ignore_index
+    # Allows passing validation specific args via transform_args:
+    # transform_args={
+    #   "image_size": ..., # train only
+    #   "normalize": ..., # train and val
+    #   "val": {
+    #       "image_size": ..., # val only
+    # }
+    val_args = transform_args.pop("val", {})
 
     train_transform_args_cls = train_model_cls.train_transform_cls.transform_args_cls
     val_transform_args_cls = train_model_cls.val_transform_cls.transform_args_cls
+    train_transform_args: TaskTransformArgs
+    val_transform_args: TaskTransformArgs
 
-    if ignore_index is None:
-        return (
-            train_transform_args_cls(),
-            val_transform_args_cls(),
-        )
-
-    return train_transform_args_cls(ignore_index=ignore_index), val_transform_args_cls(  # type: ignore[call-arg]
-        ignore_index=ignore_index
+    train_transform_args = validate.pydantic_model_validate(
+        train_transform_args_cls, transform_args
     )
+    train_transform_args.resolve_auto()
+
+    # Take defaults from train transform.
+    val_args_dict = train_transform_args.model_dump(
+        include={"image_size": True, "normalize": True, "ignore_index": True}
+    )
+    # Overwrite with user provided val args.
+    val_args_dict.update(val_args)
+    val_transform_args = validate.pydantic_model_validate(
+        val_transform_args_cls, val_args_dict
+    )
+    val_transform_args.resolve_auto()
+
+    logger.debug(
+        f"Resolved train transform args {pretty_format_args(train_transform_args.model_dump())}"
+    )
+    logger.debug(
+        f"Resolved val transform args {pretty_format_args(val_transform_args.model_dump())}"
+    )
+    return train_transform_args, val_transform_args
 
 
 def get_train_transform(