+++ src/lightly_train/_task_models/dinov3_eomt_semantic_segmentation/transforms.py
@@ -7,6 +7,8 @@
 #
 from __future__ import annotations
 
+from typing import Literal
+
 from pydantic import Field
 
 from lightly_train._transforms.semantic_segmentation_transform import (
@@ -14,9 +16,8 @@ from lightly_train._transforms.semantic_segmentation_transform import (
     SemanticSegmentationTransformArgs,
 )
 from lightly_train._transforms.transform import (
-    CenterCropArgs,
+    ChannelDropArgs,
     ColorJitterArgs,
-    LongestMaxSizeArgs,
     NormalizeArgs,
     RandomCropArgs,
     RandomFlipArgs,
@@ -46,30 +47,16 @@ class DINOv3EoMTSemanticSegmentationScaleJitterArgs(ScaleJitterArgs):
 
 
 class DINOv3EoMTSemanticSegmentationSmallestMaxSizeArgs(SmallestMaxSizeArgs):
-    max_size: list[int] = [518]
+    max_size: int | list[int] | Literal["auto"] = "auto"
     prob: float = 1.0
 
 
 class DINOv3EoMTSemanticSegmentationRandomCropArgs(RandomCropArgs):
-    height: int = 518
-    width: int = 518
+    height: int | Literal["auto"] = "auto"
+    width: int | Literal["auto"] = "auto"
     pad_if_needed: bool = True
     pad_position: str = "center"
     fill: int = 0
-    prob: float = 1.0
-
-
-class DINOv3EoMTSemanticSegmentationCenterCropArgs(CenterCropArgs):
-    height: int = 518
-    width: int = 518
-    pad_if_needed: bool = True
-    pad_position: str = "center"
-    fill: int = 0
-    prob: float = 1.0
-
-
-class DINOv3EoMTSemanticSegmentationLongestMaxSizeArgs(LongestMaxSizeArgs):
-    max_size: int = 518
     prob: float = 1.0
 
 
@@ -82,6 +69,7 @@ class DINOv3EoMTSemanticSegmentationTrainTransformArgs(
 
     # TODO(Guarin, 08/25): Check if we should change default to 512.
     image_size: tuple[int, int] = (518, 518)
+    channel_drop: ChannelDropArgs | None = None
     normalize: NormalizeArgs = Field(default_factory=NormalizeArgs)
     random_flip: RandomFlipArgs = Field(default_factory=RandomFlipArgs)
     color_jitter: DINOv3EoMTSemanticSegmentationColorJitterArgs = Field(
@@ -94,8 +82,6 @@ class DINOv3EoMTSemanticSegmentationTrainTransformArgs(
     random_crop: RandomCropArgs = Field(
         default_factory=DINOv3EoMTSemanticSegmentationRandomCropArgs
     )
-    longest_max_size: LongestMaxSizeArgs | None = None
-    center_crop: CenterCropArgs | None = None
 
 
 class DINOv3EoMTSemanticSegmentationValTransformArgs(SemanticSegmentationTransformArgs):
@@ -104,6 +90,7 @@ class DINOv3EoMTSemanticSegmentationValTransformArgs(SemanticSegmentationTransfo
     """
 
     image_size: tuple[int, int] = (518, 518)
+    channel_drop: ChannelDropArgs | None = None
     normalize: NormalizeArgs = Field(default_factory=NormalizeArgs)
     random_flip: RandomFlipArgs | None = None
     color_jitter: ColorJitterArgs | None = None
@@ -112,23 +99,11 @@ class DINOv3EoMTSemanticSegmentationValTransformArgs(SemanticSegmentationTransfo
         default_factory=DINOv3EoMTSemanticSegmentationSmallestMaxSizeArgs
     )
     random_crop: RandomCropArgs | None = None
-    longest_max_size: LongestMaxSizeArgs | None = None
-    center_crop: CenterCropArgs | None = None
 
 
 class DINOv3EoMTSemanticSegmentationTrainTransform(SemanticSegmentationTransform):
     transform_args_cls = DINOv3EoMTSemanticSegmentationTrainTransformArgs
 
-    def __init__(
-        self, transform_args: DINOv3EoMTSemanticSegmentationTrainTransformArgs
-    ) -> None:
-        super().__init__(transform_args=transform_args)
-
 
 class DINOv3EoMTSemanticSegmentationValTransform(SemanticSegmentationTransform):
     transform_args_cls = DINOv3EoMTSemanticSegmentationValTransformArgs
-
-    def __init__(
-        self, transform_args: DINOv3EoMTSemanticSegmentationValTransformArgs
-    ) -> None:
-        super().__init__(transform_args=transform_args)